w5/pg-mcp/pyproject.toml
[tool.poetry]
name = "pg-mcp-server"
version = "1.0.0"
description = "PostgreSQL MCP Server - Natural Language to SQL Query System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pg_mcp_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastmcp = "^0.1.0"
asyncpg = "^0.31.0"
sqlglot = "^20.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
openai = "^1.10.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
structlog = "^24.0.0"
tenacity = "^8.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
pg-mcp-server = "pg_mcp_server.__main__:main"
pg-mcp = "pg_mcp_server.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
